requirements.txt
Flask==3.1.2
gradio==5.44.1
MarkupSafe==3.0.2
openai==1.102.0
pandas==2.3.2
python-dotenv==1.1.1
PyMySQL==1.1.1
mysql-connector-python==9.1.0
Pillow==10.0.1
pymongo==4.6.1
